Updated Dockerfile and removed yarnrc (#5)
* Updated Dockerfile and removed yarnrc

Bumped the nginx-alpine image to the latest stable. We should always use
at least stable if not mainline and auto-update regularly.

The version of yarn in yarnrc was no longer available causing the build
to fail. Removing that file allowed the latest patch of 1.22.x to be
installed. I am not 100% sure this is the correct approach to managing
yarn in the container, but it works for now and the build is successful.

* Trigger Build